+++ src/main/java/wooteco/subway/admin/domain/CriteriaType.java
@@ -4,21 +4,18 @@ import java.util.Arrays;
 import java.util.function.Function;
 
 public enum CriteriaType {
-    DISTANCE("distance", LineStation::getDistance),
-    DURATION("duration", LineStation::getDuration);
+    DISTANCE(LineStation::getDistance),
+    DURATION(LineStation::getDuration);
 
-    private final String name;
     private final Function<LineStation, Integer> expression;
 
-    CriteriaType(String name,
-        Function<LineStation, Integer> expression) {
-        this.name = name;
+    CriteriaType(Function<LineStation, Integer> expression) {
         this.expression = expression;
     }
 
     public static CriteriaType of(String input) {
         return Arrays.stream(values())
-            .filter(instance -> instance.name.equalsIgnoreCase(input))
+            .filter(instance -> instance.name().equalsIgnoreCase(input))
             .findFirst()
             .orElseThrow(AssertionError::new);
     }
